🎨 Séparer specs tâche planifiée
