+++ packages/specifications/src/tâche/stepDefinitions/tâche.then.ts
@@ -4,10 +4,9 @@ import waitForExpect from 'wait-for-expect';
 import { expect } from 'chai';
 
 import { ListerTâchesQuery } from '@potentiel-domain/tache';
-import { ListerTâchesPlanifiéesQuery } from '@potentiel-domain/tache-planifiee';
 
 import { PotentielWorld } from '../../potentiel.world';
-import { RechercherTypeTâche, RechercherTypeTâchePlanifiée } from '../tâche.world';
+import { RechercherTypeTâche } from '../tâche.world';
 
 Alors(
   `une tâche indiquant de {string} est consultable dans la liste des tâches du porteur pour le projet`,
@@ -43,34 +42,6 @@ Alors(
 
       const tâche = tâches.items.find((t) => t.typeTâche.estÉgaleÀ(actualTypeTâche));
       expect(tâche).to.be.undefined;
-    });
-  },
-);
-
-Alors(
-  `une tâche {string} est planifiée à la date du {string} pour le projet {string}`,
-  async function (
-    this: PotentielWorld,
-    typeTâche: RechercherTypeTâchePlanifiée,
-    dateTâche: string,
-    nomProjet: string,
-  ) {
-    const actualTypeTâche = this.tâcheWorld.rechercherTypeTâchePlanifiée(typeTâche);
-    const projet = this.lauréatWorld.rechercherLauréatFixture(nomProjet);
-
-    await waitForExpect(async () => {
-      const tâches = await mediator.send<ListerTâchesPlanifiéesQuery>({
-        type: 'Tâche.Query.ListerTâchesPlanifiées',
-        data: {
-          àExécuterLe: new Date(dateTâche).toISOString(),
-        },
-      });
-      const tâche = tâches.items.find(
-        (t) =>
-          t.typeTâchePlanifiée.estÉgaleÀ(actualTypeTâche) &&
-          t.identifiantProjet.estÉgaleÀ(projet.identifiantProjet),
-      );
-      expect(tâche).not.to.be.undefined;
     });
   },
 );
