+++ packages/specifications/src/tâche/tâche.world.ts
@@ -1,13 +1,10 @@
 import { TypeTâche } from '@potentiel-domain/tache';
-import { TypeTâchePlanifiée } from '@potentiel-domain/tache-planifiee';
 
 export type RechercherTypeTâche =
   | 'transmettre la preuve de recandidature'
   | 'confirmer un abandon'
   | 'transmettre les garanties financières'
   | 'mettre à jour le gestionnaire de réseau';
-
-export type RechercherTypeTâchePlanifiée = 'échoir les garanties financières';
 
 export class TâcheWorld {
   rechercherTypeTâche(value: RechercherTypeTâche): TypeTâche.ValueType {
@@ -22,15 +19,6 @@ export class TâcheWorld {
         return TypeTâche.raccordementGestionnaireRéseauInconnuAttribué;
       default:
         return TypeTâche.inconnue;
-    }
-  }
-
-  rechercherTypeTâchePlanifiée(value: RechercherTypeTâchePlanifiée): TypeTâchePlanifiée.ValueType {
-    switch (value) {
-      case 'échoir les garanties financières':
-        return TypeTâchePlanifiée.garantiesFinancieresÉchoir;
-      default:
-        return TypeTâchePlanifiée.inconnue;
     }
   }
 }
